trekk SykmeldingHendelse til sin egen fil
