+++ src/main/kotlin/no/nav/helse/flex/sykmelding/domain/Sykmelding.kt
@@ -46,21 +46,4 @@ data class Sykmelding(
         )
 }
 
-data class SykmeldingHendelse(
-    internal val databaseId: String? = null,
-    val status: HendelseStatus,
-    val sporsmalSvar: List<Sporsmal>? = null,
-    val arbeidstakerInfo: ArbeidstakerInfo? = null,
-    val opprettet: Instant,
-)
-
-enum class HendelseStatus {
-    APEN,
-    AVBRUTT,
-    SENDT_TIL_NAV,
-    SENDT_TIL_ARBEIDSGIVER,
-    BEKREFTET_AVVIST,
-    UTGATT,
-}
-
 private fun <T, R : Comparable<R>> Iterable<T>.lastMaxBy(selector: (T) -> R): T? = this.sortedBy(selector).lastOrNull()
